Added position to Note. Made me rework and restructure the service layer. Also change up the requests from the frontend

+++ src/main/java/com/company/projectManager/whiteboard/columns/dto/ColumnDTO.java
@@ -14,7 +14,7 @@ public record ColumnDTO(
         WhiteboardDTO whiteboardDTO,
 
         @NotNull
-        int position
+        Long position
 ) {
 
 }
